% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{crps}
\alias{crps}
\title{Calculates CRPS}
\usage{
crps(y, mu, s2)
}
\arguments{
\item{y}{response vector}

\item{mu}{predicted mean}

\item{s2}{predicted point-wise variances}
}
\description{
Calculates continuous ranked probability score (lower CRPS indicate
better fits, better uncertainty quantification).
}
\examples{
# Additional examples including real-world computer experiments are available at: 
# https://bitbucket.org/gramacylab/deepgp-ex/

}
\references{
Gneiting, T, and AE Raftery. 2007. Strictly Proper Scoring Rules, Prediction, 
    and Estimation. \emph{Journal of the American Statistical Association 102} 
    (477), 359-378.
}
