% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rmse}
\alias{rmse}
\title{Calculates RMSE}
\usage{
rmse(y, mu)
}
\arguments{
\item{y}{response vector}

\item{mu}{predicted mean}
}
\description{
Calculates root mean square error (lower RMSE indicate better 
    fits).
}
\examples{
# Additional examples including real-world computer experiments are available at: 
# https://bitbucket.org/gramacylab/deepgp-ex/

}
