% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rmse}
\alias{rmse}
\title{Calculates RMSE}
\usage{
rmse(y, mu)
}
\arguments{
\item{y}{response vector}

\item{mu}{predicted mean}
}
\description{
Calculates root mean square error (lower RMSE indicate better fits).
}
\examples{
# See "deepgp-package", "fit_one_layer", "fit_two_layer", or "fit_three_layer"
# for an example

}
