% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{score}
\alias{score}
\title{Calculates score}
\usage{
score(y, mu, sigma)
}
\arguments{
\item{y}{response vector}

\item{mu}{predicted mean}

\item{sigma}{predicted covariance}
}
\description{
Calculates score (higher scores indicate better fits).  Only 
    applicable to noisy data.  Requires full covariance matrix (e.g. 
    \code{predict} with \code{lite = FALSE}).
}
\references{
Gneiting, T, and AE Raftery. 2007. Strictly Proper Scoring Rules, Prediction, 
    and Estimation. \emph{Journal of the American Statistical Association 
    102} (477), 359-378.
}
